JALUserDataTypeEnum is going into retirement
I'm going to scrap that whole enum/array thing because although it's a
perfectly good technique in come situations, in this case, where I've
only got 2 strings related to each "thing" and not only that but
currently only 1 of those "things" (user data type) I can just use
normal consts! Why so complicated!!!

I'm going to retire it into a branch though because a lot of brain
effort went into it, regarding testing. It'd be a shame to lose that if
I need to reintroduce sth like this.